create tests fixed, compilers and tools refactor
